+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Change Wine GE version from Wine-GE-Proton7-43 to Wine-GE-Proton8-17.
 - Change GE Proton version from GE-Proton7-55 to GE-Proton8-16.
 - Add the Nix package manager.
+- Add Gamescope Session.
 
 ## 3.3.0 - The Major Upgrade Update